Move SDKConfig to the edge subpackage
